requirements.txt
Django == 1.8
mongoengine == 0.10.0
djangorestframework == 3.*
mock == 1.0.1
qualname == 0.1.0